move navbar into common folder
