+++ components/footer/footerItem/footerItem.js
@@ -1,5 +1,5 @@
 import clsx from "clsx";
-import NavbarLink from "../../navbarLink/navbarLink";
+import NavbarLink from "../../navbar/navbarLink/navbarLink";
 import styles from "../footer/footer.module.scss";
 import utilStyles from "../../styles/utils.module.scss";
