+++ components/navbar/navbarItem/navbarItem.js
@@ -1,6 +1,6 @@
 import EscapeOutside from "react-escape-outside";
 import NavbarLink from "../navbarLink/navbarLink";
-import NavbarMenu from "../navbarMenu/navbarMenu";
+import NavbarMenu from "../../navbarMenu/navbarMenu";
 import { useRouter } from "next/router";
 import { useState, useEffect } from "react";
 import styles from "../navbar/navbar.module.scss";
